requirements.txt
flask~=2.2.5
flask_cors~=4.0.0
deepface~=0.0.79
gunicorn~=21.2.0
Pillow==10.0.1